simplify the constraint implementation
The constraint implementation was suffering from a design problem by which
functions could only be called in a certain order. This was caused by
Constraint being a hibrid between a builder and an actual object.

Builders are supposed to check for the correctness of the configuration
only at the end when the `build` function is called and when all the
configuration is set. The constraint implementation though was checking
the alignment when it was set. When the allocation policy is Exact Match,
the start address MUST also be aligned. If the caller would first call
`set_align` and then `set_policy` with ExactMatch, than the caller
would be allowed to specify an unaligned address.

There are 2 possible solutions to fix this problem:
1. Create a ConstraintBuilder separate from Constraint -> this results
in code bloat, and for such a simple structure is superflous.
2. Initialize all the configuration for the Constraint in one go in the
constructor -> the disadvantage with this one is that both align and size
have the type u64, and users of the library might mistakenly switch the
two when calling the function. With IDEs these days, this concern should
be alliviated.

Fixes: https://github.com/rust-vmm/vm-allocator/issues/29

Signed-off-by: Andreea Florescu <fandree@amazon.com>