updated readme
The readme was not pointing to the interval tree design details. Added
a link so that it's easier for users to find this documentation.

Split the README.md & DESIGN.md at 80 chars.

Updated some of the outdated information, and make it less prone to
outdates by removing code blocks and favoring using a description of
structures instead of the structure names (when this is possible).

Signed-off-by: Andreea Florescu <fandree@amazon.com>